dom/tests/mochitest/general/test_CCW_optimization.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1319087
-->
<head>
  <title>Test for Bug 1319087</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1319087">Mozilla Bug 1319087</a>
<p id="display"></p>
<div id="content">
  <iframe></iframe>
  <iframe></iframe>
</div>
<pre id="test">
<script type="text/javascript">

function WrapperToOwnCompartment() {
  var iframe = new frames[0].Object();
  var obj = iframe.obj = new Object();
  obj.x = 123;

  for (var i = 0; i < 50; i++) {
    is(iframe.obj, obj);
    is(iframe.obj.x, 123);
  }
}

function WrapperToYetAnotherCompartment() {
  var iframe = new frames[0].Object();
  // Obj points to an object in a third compartment.
  var obj = iframe.obj = new frames[1].Object();
  obj.x = 42;

  for (var i = 0; i < 50; i++) {
    is(iframe.obj, obj);
    is(iframe.obj.x, 42);
  }
}

WrapperToOwnCompartment();
WrapperToYetAnotherCompartment();

</script>
</body>
</html>
